got rid of warnings on windows, removed include of c++ string library

+++ src/game_platform_interface.h
@@ -84,13 +84,13 @@ struct GameInputBuffer
 
 
 // debug/prototyping functions only
-#define FUNC_DEBUG_PLATFORM_READ_ENTIRE_FILE(name) void* name(const char* filename, int32_t* returned_size)
+#define FUNC_DEBUG_PLATFORM_READ_ENTIRE_FILE(name) void* name(const char* filename, int64_t* returned_size)
 typedef FUNC_DEBUG_PLATFORM_READ_ENTIRE_FILE(DEBUGPlatformReadEntireFile);
 
 #define FUNC_DEBUG_PLATFORM_FREE_FILE_MEMORY(name) void name(void* memory)
 typedef FUNC_DEBUG_PLATFORM_FREE_FILE_MEMORY(DEBUGPlatformFreeFileMemory);
 
-#define FUNC_DEBUG_PLATFORM_WRITE_ENTIRE_FILE(name) void name(const char* filename, void* buffer, int32_t len)
+#define FUNC_DEBUG_PLATFORM_WRITE_ENTIRE_FILE(name) void name(const char* filename, void* buffer, int64_t len)
 typedef FUNC_DEBUG_PLATFORM_WRITE_ENTIRE_FILE(DEBUGPlatformWriteEntireFile);
 //
